requirements.txt
flask>=2.0.0
requests>=2.25.0
streamlit>=1.0.0
pandas>=1.3.0
plotly>=5.0.0
python-dotenv>=0.19.0
numpy>=1.21.0